Added router successfully
