+++ client/package.json
@@ -6,6 +6,7 @@
     "js-confetti": "^0.12.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.0",
     "react-scripts": "5.0.1"
   },
   "scripts": {
