+++ client/src/Statistics.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Statistics() {
+  return (
+    <h1>Statistics</h1>
+  );
+}
+
+export default Statistics;
